+++ src/main/kotlin/io/verik/intellij/arguments/VerikArgumentsHelperUtil.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2022 Francis Wang
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.verik.intellij.arguments
+
+import org.jetbrains.kotlin.descriptors.CallableDescriptor
+import org.jetbrains.kotlin.descriptors.ClassConstructorDescriptor
+import org.jetbrains.kotlin.descriptors.SimpleFunctionDescriptor
+import org.jetbrains.kotlin.idea.caches.resolve.resolveToCall
+import org.jetbrains.kotlin.psi.KtCallExpression
+import org.jetbrains.kotlin.psi.KtValueArgumentList
+import org.jetbrains.kotlin.psi.psiUtil.getStrictParentOfType
+
+object VerikArgumentsHelperUtil {
+
+    fun getDescriptor(list: KtValueArgumentList): CallableDescriptor? {
+        val calleeExpression = list.getStrictParentOfType<KtCallExpression>()?.calleeExpression ?: return null
+        return when (val descriptor = calleeExpression.resolveToCall()?.resultingDescriptor) {
+            is ClassConstructorDescriptor -> descriptor
+            is SimpleFunctionDescriptor -> descriptor
+            else -> null
+        }
+    }
+}