+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Highlight sequential assignments.
+- Inspection to add missing arguments.
 
 ## [0.1.4]
 ### Added